backend/requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ─── Validation & config ──────────────────────────────────────────────────────
pydantic>=2.9,<3
pydantic-settings>=2.6,<3

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.36,<3
asyncpg==0.30.0
alembic==1.14.0

# ─── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
httpx==0.27.2

# ─── LLM / Agent ──────────────────────────────────────────────────────────────
anthropic>=0.40.0
langgraph>=0.2.50
langchain-anthropic>=0.3.0
langchain-core>=0.3.0

# ─── RAG (Phase 2) ────────────────────────────────────────────────────────────
weaviate-client>=4.9.0
tiktoken>=0.8.0
pdfplumber>=0.11.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# ─── Tools (Phase 3) ──────────────────────────────────────────────────────────
sqlparse>=0.5.0
cachetools>=5.5.0

# ─── Observability ────────────────────────────────────────────────────────────
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-instrumentation-fastapi>=0.50b0
opentelemetry-instrumentation-sqlalchemy>=0.50b0
opentelemetry-instrumentation-httpx>=0.50b0
opentelemetry-exporter-otlp>=1.29.0

# ─── Logging ──────────────────────────────────────────────────────────────────
structlog>=24.4.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-multipart==0.0.12
